+++ ItaCH_Smash_Legends/Assets/Script/Alice/AliceEffectController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8e9ced9b6f18f04695fd4333fc85efb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
