+++ ItaCH_Smash_Legends/Assets/Script/Player_Effect/EffectController.cs
@@ -30,12 +30,4 @@ public abstract class EffectController : MonoBehaviour
             _effects[i].SetActive(false);
         }
     }
-    public abstract void EnableFirstAttackEffect();
-    public abstract void EnableSecondAttackFirstEffect();
-    public abstract void EnableSecondAttackSecondEffect();
-    public abstract void EnableFinishAttackEffect();
-    public abstract void EnableHeavyAttackEffect();
-    public abstract void EnableJumpAttackEffect();
-    public abstract void DisableJumpAttackEffect();
-
 }
